requirements.txt
streamlit
requests
beautifulsoup4
praw
python-dotenv
